change coupon discount type

+++ db/migrate/20221214144833_create_coupons.rb
@@ -1,7 +1,7 @@
 class CreateCoupons < ActiveRecord::Migration[6.1]
   def change
     create_table :coupons do |t|
-      t.float :discount, null: false
+      t.integer :discount, null: false
       t.string :code, null: false
       t.references :cart, foreign_key: true, index: { unique: true }
